reorganize bigly structure of tests
