+++ tests/data/pruned_comps.tsv
@@ -1,0 +1,10 @@
+ID	Type	Generation	Formula	InChIKey	SMILES
+ADP	Coreactant	0	C10H15N5O10P2	XTWYTFMLZFPYCI-KQYNXXCUSA-N	Nc1ncnc2c1ncn2[C@@H]1O[C@H](COP(=O)(O)OP(=O)(O)O)[C@@H](O)[C@H]1O
+ATP	Coreactant	0	C10H16N5O13P3	ZKHQWZAMYRWXGA-KQYNXXCUSA-N	Nc1ncnc2c1ncn2[C@@H]1O[C@H](COP(=O)(O)OP(=O)(O)OP(=O)(O)O)[C@@H](O)[C@H]1O
+FADH	Starting Compound	0	C27H35N9O15P2	YPZRHBJKEMOYQH-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(O)C1O)c1nc(O)nc(O)c1N2
+pkc0000001	Predicted	1	C27H36N9O18P3	WZDVFMUFGAERLJ-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(OP(=O)(O)O)C1O)c1nc(O)nc(O)c1N2
+pkc0000002	Predicted	1	C27H36N9O18P3	APYAYBLNPQXBSR-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(O)C1O)c1nc(O)nc(OP(=O)(O)O)c1N2
+pkc0000003	Predicted	1	C27H36N9O18P3	IDFBYYNFURDRLI-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(OP(=O)(O)O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(O)C1O)c1nc(O)nc(O)c1N2
+pkc0000004	Predicted	1	C27H36N9O18P3	IRPOHFRNKHKIQA-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(O)C1OP(=O)(O)O)c1nc(O)nc(O)c1N2
+pkc0000005	Predicted	2	C27H37N9O21P4	AOMDRBXRIDYZPG-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(OP(=O)(O)O)C1OP(=O)(O)O)c1nc(O)nc(O)c1N2
+pkc0000006	Predicted	2	C27H37N9O21P4	MJMURRAVZDFEFB-UHFFFAOYSA-N	Cc1cc2c(cc1C)N(CC(OP(=O)(O)O)C(O)C(O)COP(=O)(O)OP(=O)(O)OCC1OC(n3cnc4c(N)ncnc43)C(O)C1O)c1nc(O)nc(OP(=O)(O)O)c1N2
