+++ tests/test_unit/test_metabolomics.py
@@ -1,6 +1,7 @@
 """Tests for metabolomics.py using pytest."""
 
 import os
+from pathlib import Path
 
 import pymongo
 import pytest
@@ -20,12 +21,18 @@ from minedatabase.metabolomics import (
 
 
 try:
-    client = pymongo.MongoClient(ServerSelectionTimeoutMS=2000)
+    client = pymongo.MongoClient(ServerSelectionTimeoutMS=20)
+    client.server_info()
     del client
     is_mongo = True
 except ServerSelectionTimeoutError as err:
     is_mongo = False
 valid_db = pytest.mark.skipif(not is_mongo, reason="No MongoDB Connection")
+
+file_path = Path(__file__)
+file_dir = file_path.parent
+
+DATA_DIR = (file_dir / "../data/").resolve()
 
 # -------------------------------- Fixtures --------------------------------- #
 @pytest.fixture()
@@ -279,7 +286,7 @@ def test_read_mgf():
     WHEN that MGF file is parsed into a list of Peak objects
     THEN make sure those Peak objects are correct
     """
-    mgf_path = os.path.join(os.path.dirname(__file__), "data/metabolomics/test.mgf")
+    mgf_path = DATA_DIR / "test_metabolomics/test.mgf"
     with open(mgf_path, "r") as infile:
         mgf_data = infile.read()
 
@@ -304,7 +311,7 @@ def test_read_msp():
     WHEN that MSP file is parsed into a list of Peak objects
     THEN make sure those Peak objects are correct
     """
-    msp_path = os.path.join(os.path.dirname(__file__), "data/metabolomics/test.msp")
+    msp_path = DATA_DIR / "test_metabolomics/test.msp"
     with open(msp_path, "r") as infile:
         msp_data = infile.read()
 
@@ -329,7 +336,7 @@ def test_read_mzxml():
     WHEN that mzXML file is parsed into a list of Peak objects
     THEN make sure those Peak objects are correct
     """
-    mzxml_path = os.path.join(os.path.dirname(__file__), "data/metabolomics/test.mzXML")
+    mzxml_path = DATA_DIR / "test_metabolomics/test.mzXML"
     with open(mzxml_path, "r") as infile:
         mzxml_data = infile.read()
 
@@ -375,9 +382,8 @@ def test_spectra_download(test_db):
     cpd_query = '{"_id": "ms2_test"}'
     spectra = spectra_download(test_db, cpd_query)
 
-    spectra_path = os.path.join(
-        os.path.dirname(__file__), "data/metabolomics/test_spectra.txt"
-    )
+    spectra_path = DATA_DIR / "test_metabolomics/test_spectra.txt"
+
     with open(spectra_path, "r") as infile:
         spectra_data = infile.read()
